# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Embed Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Folder {
  id: number;
  name: string;
  /** @nullable */
  parentId?: number | null;
  createdAt: string;
  embedCount: number;
}

export interface FolderInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  parentId?: number | null;
}

export interface FolderUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  parentId?: number | null;
}

export interface Embed {
  id: number;
  title: string;
  /** @nullable */
  url?: string | null;
  embedCode: string;
  /** @nullable */
  thumbnail?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  folderName?: string | null;
  createdAt: string;
}

export interface EmbedInput {
  /** @minLength 1 */
  title: string;
  url?: string;
  /** @minLength 1 */
  embedCode: string;
  thumbnail?: string;
  description?: string;
  source?: string;
  /** @nullable */
  folderId?: number | null;
}

export interface EmbedUpdate {
  /** @minLength 1 */
  title?: string;
  url?: string;
  embedCode?: string;
  thumbnail?: string;
  description?: string;
  source?: string;
  /** @nullable */
  folderId?: number | null;
}

export interface BulkIds {
  ids: number[];
}

export interface BulkMove {
  ids: number[];
  /** @nullable */
  folderId?: number | null;
}

export interface BulkResult {
  count: number;
}

export interface Stats {
  totalEmbeds: number;
  totalFolders: number;
  unorganizedCount: number;
  recentEmbeds: Embed[];
}

export type ListEmbedsParams = {
/**
 * @nullable
 */
folderId?: number | null;
/**
 * @nullable
 */
unorganized?: boolean | null;
};