# lib/api-zod/src/generated/types/bulkMove.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Embed Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface BulkMove {
  ids: number[];
  /** @nullable */
  folderId?: number | null;
}