# lib/api-zod/src/generated/types/embed.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Embed Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Embed {
  id: number;
  title: string;
  /** @nullable */
  url?: string | null;
  embedCode: string;
  /** @nullable */
  thumbnail?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  folderName?: string | null;
  createdAt: string;
}