# lib/api-zod/src/generated/types/embedUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Embed Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface EmbedUpdate {
  /** @minLength 1 */
  title?: string;
  url?: string;
  embedCode?: string;
  thumbnail?: string;
  description?: string;
  source?: string;
  /** @nullable */
  folderId?: number | null;
}