# lib/api-zod/src/generated/types/folder.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Embed Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Folder {
  id: number;
  name: string;
  /** @nullable */
  parentId?: number | null;
  createdAt: string;
  embedCount: number;
}