# lib/api-zod/src/generated/types/folderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Embed Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface FolderInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  parentId?: number | null;
}