# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Embed Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Embed } from './embed';

export interface Stats {
  totalEmbeds: number;
  totalFolders: number;
  unorganizedCount: number;
  recentEmbeds: Embed[];
}